# influxdata: add benchmark for top level tokenizing methods