--- a/influxdata/byteset.go
+++ b/influxdata/byteset.go
@@ -10,25 +10,23 @@ func newByteSet(s string) *byteSet {
 	return &set
 }
 
-// TODO benchmark it. This is compact (good cache behaviour)
-// but maybe [256]bool might be faster (less operations).
-type byteSet [4]uint64
+type byteSet [256]bool
 
-// get reports whether b holds the byte x.
+// holds reports whether b holds the byte x.
 func (b *byteSet) get(x uint8) bool {
-	return b[x>>6]&(1<<(x&63)) != 0
+	return b[x]
 }
 
 // set ensures that x is in the set.
 func (b *byteSet) set(x uint8) {
-	b[x>>6] |= 1 << (x & 63)
+	b[x] = true
 }
 
 // union returns the union of b and b1.
 func (b *byteSet) union(b1 *byteSet) *byteSet {
 	r := *b
 	for i := range r {
-		r[i] |= b1[i]
+		r[i] = r[i] || b1[i]
 	}
 	return &r
 }
@@ -37,7 +35,7 @@ func (b *byteSet) union(b1 *byteSet) *byteSet {
 func (b *byteSet) invert() *byteSet {
 	r := *b
 	for i := range r {
-		r[i] = ^r[i]
+		r[i] = !r[i]
 	}
 	return &r
 }