Also change the representation of byte sets to `[256]bool` which
speeds things up by quite a bit.

Here's the benchmark comparison between the version that uses
`[4]uint64` and the version proposed in this PR.

```
name                                                         old time/op    new time/op     delta
ScanEntriesSkipping/long-lines-8                               36.2ms ± 5%     16.7ms ± 2%   -53.81%  (p=0.008 n=5+5)
ScanEntriesSkipping/long-lines-with-escapes-8                  21.7ms ± 1%     17.8ms ± 1%   -18.20%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-8                         288ns ± 6%      237ns ±12%   -17.71%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-with-escapes-8            303ns ± 3%      228ns ± 4%   -24.77%  (p=0.008 n=5+5)
ScanEntriesSkipping/many-short-lines-8                          143ms ± 3%      130ms ±18%      ~     (p=0.151 n=5+5)
ScanEntriesSkipping/field-key-escape-not-escapable-8            271ns ± 2%      278ns ±14%      ~     (p=0.508 n=5+5)
ScanEntriesSkipping/tag-value-triple-escape-space-8             375ns ±14%      278ns ± 6%   -25.80%  (p=0.008 n=5+5)
ScanEntriesSkipping/procstat-8                                 3.83µs ± 2%     2.26µs ± 1%   -41.05%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-8                        42.9ms ± 1%     18.4ms ± 1%   -57.01%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-with-escapes-8            116ms ± 5%       93ms ± 1%   -19.38%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-8                  392ns ± 2%      304ns ± 3%   -22.45%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-with-escapes-8     462ns ±26%      323ns ± 1%   -30.07%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/many-short-lines-8                   243ms ±16%      153ms ± 1%   -37.02%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/field-key-escape-not-escapable-8     400ns ± 8%      282ns ± 3%   -29.50%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/tag-value-triple-escape-space-8      478ns ± 7%      379ns ±10%   -20.79%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/procstat-8                          7.02µs ± 7%     5.23µs ± 2%   -25.51%  (p=0.008 n=5+5)
Tokenize-8                                                     43.4ms ±10%     20.1ms ± 3%   -53.83%  (p=0.008 n=5+5)

name                                                         old speed      new speed       delta
ScanEntriesSkipping/long-lines-8                              725MB/s ± 5%   1568MB/s ± 2%  +116.25%  (p=0.008 n=5+5)
ScanEntriesSkipping/long-lines-with-escapes-8                1.21GB/s ± 1%   1.48GB/s ± 1%   +22.24%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-8                       101MB/s ± 6%    123MB/s ±11%   +22.11%  (p=0.008 n=5+5)
ScanEntriesSkipping/single-short-line-with-escapes-8          106MB/s ± 3%    141MB/s ± 4%   +32.99%  (p=0.008 n=5+5)
ScanEntriesSkipping/many-short-lines-8                        183MB/s ± 3%    203MB/s ±16%      ~     (p=0.151 n=5+5)
ScanEntriesSkipping/field-key-escape-not-escapable-8          122MB/s ± 2%    120MB/s ±15%      ~     (p=0.548 n=5+5)
ScanEntriesSkipping/tag-value-triple-escape-space-8           134MB/s ±16%    180MB/s ± 6%   +33.84%  (p=0.008 n=5+5)
ScanEntriesSkipping/procstat-8                                347MB/s ± 2%    588MB/s ± 1%   +69.62%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-8                       612MB/s ± 1%   1423MB/s ± 1%  +132.61%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/long-lines-with-escapes-8          227MB/s ± 5%    281MB/s ± 1%   +23.96%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-8               74.0MB/s ± 2%   95.4MB/s ± 3%   +28.93%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/single-short-line-with-escapes-8  70.4MB/s ±22%   99.2MB/s ± 1%   +40.92%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/many-short-lines-8                 109MB/s ±15%    171MB/s ± 1%   +56.85%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/field-key-escape-not-escapable-8  82.7MB/s ± 9%  117.1MB/s ± 3%   +41.56%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/tag-value-triple-escape-space-8    105MB/s ± 7%    133MB/s ±11%   +26.67%  (p=0.008 n=5+5)
ScanEntriesWithoutSkipping/procstat-8                         189MB/s ± 8%    254MB/s ± 2%   +33.93%  (p=0.008 n=5+5)
Tokenize-8                                                    606MB/s ±10%   1308MB/s ± 3%  +115.90%  (p=0.008 n=5+5)
```